# apps/2-api-server/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Web Evolution API
  version: 0.1.0
  description: Shared REST backend for products and categories
servers:
  - url: http://localhost:3001
paths:
  /api/products:
    get:
      summary: List products with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: integer
      responses:
        "200":
          description: Paginated product data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProducts"
    post:
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /api/products/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
    get:
      summary: Get a single product
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a product
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a product
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/categories:
    get:
      summary: List categories
      responses:
        "200":
          description: All categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /api/categories/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
    get:
      summary: Get category with products
      responses:
        "200":
          description: Category data with products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithProducts"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update category
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete category
      responses:
        "200":
          description: Deleted category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
        categoryId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, quantity, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
        categoryId:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
        categoryId:
          type: integer
    PaginatedProducts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        parentId:
          type: integer
          nullable: true
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        parentId:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        parentId:
          type: integer
          nullable: true
    CategoryWithProducts:
      allOf:
        - $ref: "#/components/schemas/Category"
        - type: object
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: integer
              message:
                type: string